Showing the player's cards in hand; Future Plans
been researching on how to show the hand of the player with their cards by moving the cursor at the bottom of the screen. The logic for the cards to show makes since but the wrong attribute was used.

I used blit instead of sprites when working with this and found that sprites are the ideal way to go when working with a card game like this. Regardless its been interesting getting the UI working with the images! What's still needed is:

Dragging and Dropping of the Cards
Button to End the Turn
Waiting for Opponent Screen